bonjour: cast to long before converting a NSInteger to a string
As advised in the Apple documentation [^1].

[^1]: https://developer.apple.com/library/archive/documentation/CoreFoundation/Conceptual/CFStrings/formatSpecifiers.html#//apple_ref/doc/uid/TP40004265

+++ modules/services_discovery/bonjour.m
@@ -351,7 +351,7 @@ static NSString * ipAddressAsStringForData(NSData * data)
 {
     vlc_renderer_discovery_t *p_rd = (vlc_renderer_discovery_t *)_p_this;
 
-    NSString *uri = [NSString stringWithFormat:@"%@://%@:%ld", protocol, netService.hostName, netService.port];
+    NSString *uri = [NSString stringWithFormat:@"%@://%@:%ld", protocol, netService.hostName, (long)netService.port];
     NSDictionary *txtDict = [NSNetService dictionaryFromTXTRecordData:[netService TXTRecordData]];
     NSString *displayName = netService.name;
     int rendererFlags = 0;
@@ -414,7 +414,7 @@ static NSString * ipAddressAsStringForData(NSData * data)
     if ([protocol isEqualToString:@"smb"]) {
         host = ipAddressAsStringForData(netService.addresses.firstObject);
     }
-    NSString *uri = [NSString stringWithFormat:@"%@://%@:%ld", protocol, host, netService.port];
+    NSString *uri = [NSString stringWithFormat:@"%@://%@:%ld", protocol, host, (long)netService.port];
 
     input_item_t *p_input_item = input_item_NewDirectory([uri UTF8String], [netService.name UTF8String], ITEM_NET );
     if (p_input_item != NULL) {
